ambuj-legal-llm-starter/requirements.txt
gradio>=4.0.0
llama-cpp-python>=0.2.90
python-dotenv>=1.0.1
fastapi>=0.115.0
huggingface-hub>=0.24.0
uvicorn[standard]>=0.30.0
